requirements.txt
click==6.6
pluggy==0.3.1
py==1.4.31
pytest==2.9.2
PyYAML==3.11
tox==2.3.1
virtualenv==15.0.2
